move cookie utils to server only
