+++ app/lib/coveo.engine.ts
@@ -1,5 +1,4 @@
 import {
-  type CommerceSearchParameters,
   defineCommerceEngine,
   defineFacetGenerator,
   definePagination,
@@ -20,12 +19,7 @@ import {
   defineRecommendations,
   CommerceEngineDefinitionOptions,
 } from '@coveo/headless-react/ssr-commerce';
-import type {AppLoadContext} from '@shopify/remix-oxygen';
-import {getLocaleFromRequest} from './i18n';
-import type {CartReturn} from '@shopify/hydrogen';
-import {mapShopifyMerchandiseToCoveoCartItem} from './map.coveo.shopify';
 import {fetchToken} from './fetch-token';
-import {updateTokenIfNeeded} from '~/lib/token-utils';
 
 // Headless requires an `accessToken` to be set in the configuration.
 // We can't simply call `fetchToken` in all cases, because when the file
@@ -44,7 +38,7 @@ const getPublicApiKey = () => {
 
 export const engineConfig: CommerceEngineDefinitionOptions = {
   configuration: {
-    accessToken: getPublicApiKey(),
+    accessToken: await getSearchToken(),
     organizationId: 'barcagroupproductionkwvdy6lp',
     analytics: {
       enabled: true,
@@ -142,109 +136,3 @@ export type StandaloneStaticState = InferStaticState<
 export type StandaloneHydratedState = InferHydratedState<
   typeof standaloneEngineDefinition
 >;
-
-export async function fetchStaticState({
-  k,
-  parameters,
-  url,
-  context,
-  request,
-}: {
-  k:
-    | 'listingEngineDefinition'
-    | 'searchEngineDefinition'
-    | 'standaloneEngineDefinition';
-  parameters: CommerceSearchParameters;
-  url: string;
-  context: AppLoadContext;
-  request: Request;
-}) {
-  const {country, language, currency} = getLocaleFromRequest(request);
-
-  const cart = await context.cart.get();
-
-  await updateTokenIfNeeded(k, request)
-
-  return engineDefinition[k].fetchStaticState({
-    controllers: {
-      parameterManager: {
-        initialState: {
-          parameters,
-        },
-      },
-      cart: {
-        initialState: mapShopifyCartToCoveoCart(cart),
-      },
-      context: {
-        language: language.toLowerCase(),
-        country,
-        currency: currency as any,
-        view: {
-          url,
-        },
-      },
-    },
-  });
-}
-
-export async function fetchRecommendationStaticState({
-  k,
-  context,
-  request,
-  productId,
-}: {
-  context: AppLoadContext;
-  request: Request;
-  k: (
-    | 'homepageRecommendations'
-    | 'cartRecommendations'
-    | 'pdpRecommendationsLowerCarousel'
-    | 'pdpRecommendationsUpperCarousel'
-  )[];
-  productId?: string;
-}) {
-  const cart = await context.cart.get();
-  const {country, language, currency} = getLocaleFromRequest(request);
-
-  await updateTokenIfNeeded('recommendationEngineDefinition', request)
-
-  return engineDefinition.recommendationEngineDefinition.fetchStaticState({
-    controllers: {
-      homepageRecommendations: {
-        enabled: k.includes('homepageRecommendations'),
-        productId,
-      },
-      cartRecommendations: {
-        enabled: k.includes('cartRecommendations'),
-        productId,
-      },
-      pdpRecommendationsLowerCarousel: {
-        enabled: k.includes('pdpRecommendationsLowerCarousel'),
-        productId,
-      },
-      pdpRecommendationsUpperCarousel: {
-        enabled: k.includes('pdpRecommendationsUpperCarousel'),
-        productId,
-      },
-      cart: {
-        initialState: mapShopifyCartToCoveoCart(cart),
-      },
-      context: {
-        language: language.toLowerCase(),
-        country,
-        currency: currency as any,
-        view: {
-          url: 'https://shop.barca.group',
-        },
-      },
-    },
-  });
-}
-
-function mapShopifyCartToCoveoCart(cart: CartReturn | null) {
-  return {
-    items: cart?.lines.nodes.map((node) => {
-      return mapShopifyMerchandiseToCoveoCartItem(node);
-    }),
-  };
-}
