+++ app/routes/($locale).products.$handle.tsx
@@ -20,9 +20,9 @@ import {Sizes} from '~/components/Products/Sizes';
 import {Description} from '~/components/Products/Description';
 import {
   engineDefinition,
-  fetchRecommendationStaticState,
   useProductView,
 } from '~/lib/coveo.engine';
+import {fetchRecommendationStaticState} from '~/lib/coveo.engine.server';
 import {useCallback, useEffect, useState} from 'react';
 import {ProductRecommendations} from '~/components/Products/Recommendations';
 import {RecommendationProvider} from '~/components/Search/Context';
